School-PHP2
The project TopNews is a one-page web site made in OOPHP using MySQL database and by using MVC patterns in Laravel framework. There are registration and login on the web site. There are three roles: unauthorized user, user, and administrator. An unauthorized user can only view news, then search for news on the home page, register and log in, as well as send an email to the administrator, which is done by using the PHPMailer library. The user that is logged in can leave comments on the news, browse the gallery, search and view news by category. The administrator has an overview of all data from the database on the administration panel in the form of tables arranged by links that correspond to the tables in the database. It is possible to add, change and delete data on them. When viewing comments, the administrator can only delete them. On the page of activity administrator can make filter by date and then the result is displayed as a table on the Activities link. This was done through Ajax. A search on the page for the user, as well as on the page for the administrator, then entering comments and entering a category on the admin panel, as well as filtering by date from the table activity are made with the help of JavaScript and Ajax. Pagination is present on the home page, admin panel, as well as in the gallery. Checking the data in the forms is realized by regular expressions on the server side.
Technologies – HTML, CSS, MVC, Bootstrap, PHP, JavaScript, JQuery – vanillabox, Ajax, PHPMailer, MySQL